code_analysis_service/pyproject.toml
[tool.poetry]
name = "code-analysis-service"
version = "0.1.0"
description = "Code Analysis Service for Code Review System"
authors = ["azadehdarabi97@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
openai = "^1.12.0"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
gitpython = "^3.1.40"
celery = "^5.3.0"
redis = "^5.0.0"
ast-comments = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pylint = "^3.0.0"
mypy = "^1.8.0"
black = "^23.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.messages_control]
disable = ["C0111", "C0103"]

[tool.mypy]
ignore_missing_imports = true
strict = true 